server/requirements.txt
aniso8601==9.0.1
asttokens==2.0.8
Authlib==1.1.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.2
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pyzmq==23.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.5.0
tornado==6.2
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
